+++ app/templates/app/app.js
@@ -1,3 +1,1 @@
-(function(logger) {
-  logger.log('Hello, <%= projectName %>');
-}(console));
+angular.module('<%= projectName %>', []);