+++ app/templates/gulp/tasks/concurrent.js
@@ -1,5 +1,6 @@
 var gulp = require('gulp');
 
 gulp.task('concurrent', [
+  'test:tdd',
   'watch'
 ]);