+++ app/templates/test/config/karma.tdd.conf.js
@@ -1,0 +1,17 @@
+// Karma configuration
+// http://karma-runner.github.io/0.12/config/configuration-file.html
+
+var baseConfig = require('../config/karma.conf');
+
+module.exports = function(config) {
+  baseConfig(config);
+
+  config.set({
+    reporters: ['dots'],
+    port: 8084,
+    browsers: [
+      'PhantomJS'
+    ],
+    singleRun: false
+  });
+};