+++ app/templates/test/config/mocha-globals.js
@@ -1,0 +1,7 @@
+// use this file to place anything in the global object
+// for unit tests to use it
+
+before(function() {});
+beforeEach(function() {});
+afterEach(function() {});
+after(function() {});