+++ app/templates/test/config/protractor.chrome.conf.js
@@ -1,0 +1,2 @@
+// Get specific browser configuration.
+exports.config = require('./protractor.conf.js')('chrome');