+++ app/templates/test/specs/unit/sample.spec.js
@@ -1,0 +1,5 @@
+describe('Sample suite', function() {
+  it('should pass a simple test', function() {
+    expect(true).to.be.true;
+  });
+});