+++ app/templates/test/utils/functional-flows.js
@@ -1,0 +1,2 @@
+// Here you can place repetitive jobs for UI testing in protractor
+module.exports = {};