+++ app/templates/test/utils/utils.js
@@ -1,0 +1,2 @@
+// Here you can place helpers for testing
+module.exports = {};